+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### [14.1.0](https://kaos.sh/ek/14.1.0)
+
+- **`[req]`** `PostFile` replaced by `SendFile`
+- **`[req]`** Added method `Header.Has`
+
 ### [14.0.0](https://kaos.sh/ek/14.0.0)
 
 - **`[cache/fs]`** Added method `Stop`